Add files via upload
This is a very sparse type of byte pair encoding. It doesn't yet apply any math, but before long we will apply a number of stats to the problem of druely inferring <3 many compressed strings. It's called Dre-AM-er because depending on what book of you process using it, 'back-propagating' word, sentence, paragraph and document level weights with a perhaps random seed will generate beams which are more or less masked to oblivion, only 1.6 percent will be visible, or so. 

As a consequence of all this, BPE sampling will produce an array of most likely sentences, but these sentences will not be known to the machine nor the user, or at least very rarely known to them in absense of the relevant training data and a seed which happens to match one such data. When overfitting is avoided, the sentences are less inferable from the training data, however the BPE should have the general features which actually help decode documents which were not in the training data or are statistically rare, as very many doubtless are.